plugins/aws-agentcore-langgraph/examples/policy-demo/pyproject.toml
[project]
name = "policy-demo"
version = "0.1.0"
description = "AgentCore Policy demo - Natural language to Cedar authorization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "bedrock-agentcore-starter-toolkit",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
